Touchup usage of claims and scopes.
Jeff Gaynor pointed out that we were overloading `scopes` and `claims`
and that the two things really should be orthogonal.  Accordingly, the
authorizations are now in a claim called `authz` and I provide several
examples of how requested scopes for the authorization token should be
translated into claims in the token itself.